requirements.txt
aniso8601==8.0.0
attrs==19.3.0
Babel==2.8.0
beautifulsoup4==4.8.2
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.9.2
CurrencyConverter==0.14.1
decorator==4.4.1
Faker==2.0.4
Flask==1.1.1
Flask-RESTful==0.3.8
future==0.18.2
haikunator==2.1.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
phonenumbers==8.12.1
price-parser==0.3.1
pycparser==2.19
PyCurrency-Converter==1.0
pymongo==3.8.0
python-dateutil==2.8.1
python-http-client==3.1.0
python-telegram-bot==12.7
pytz==2019.3
requests==2.22.0
six==1.13.0
soupsieve==2.0
text-unidecode==1.3
tornado==6.0.4
ujson==2.0.3
urllib3==1.25.9
Werkzeug==1.0.1
